Create useItems file
