+++ src/front/hook/useItems.js
@@ -1,0 +1,30 @@
+const { useState, useCallback } = require('react');
+
+const useItems = () => {
+  const [items, setItems] = useState({});
+
+  const deleteChecked = useCallback(
+    /**
+     * @param {Object.<string, boolean>|undefined} checked
+     */
+    (checked) => {
+      const ret = { ...items };
+      if (checked === undefined) {
+        setItems({});
+      }
+      else {
+        for (const itemName in checked) {
+          delete ret[itemName];
+        }
+        setItems(ret);
+      }
+    }, [items, setItems]);
+
+  return {
+    items,
+    setItems,
+    deleteChecked,
+  }
+}
+
+export default useItems;
